requirements.txt
absl-py==2.3.1
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
altgraph==0.17.5
annotated-doc==0.0.4
anyio==4.13.0
appnope==0.1.4
arrow==1.4.0
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
binaryornot==0.4.4
blinker==1.9.0
bottle==0.13.4
brotlicffi==1.2.0.0
cachetools==5.5.1
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
choreographer==1.2.1
click==8.2.1
comm==0.2.3
contourpy==1.3.1
cookiecutter==2.7.1
cryptography==41.0.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.24.3
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==24.3.25
flet==0.84.0
flet-cli==0.84.0
flet-desktop==0.84.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2026.2.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.47.0
google-auth-oauthlib==0.5.2
google-pasta==0.2.0
grad-cam==1.5.5
grpcio==1.48.2
h11==0.16.0
h5py==3.12.1
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kaleido==1.2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
logistro==2.0.1
macholib==1.16.4
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
narwhals==2.18.0
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
oauthlib==3.3.1
opencv-python==4.13.0.92
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.8
packaging==25.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
piexif==1.1.3
pillow==11.1.0
pip==26.0.1
platformdirs==4.9.2
plotly==6.6.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==3.20.3
proxy_tools==0.1.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycares==4.10.0
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
pyOpenSSL==23.2.0
pyparsing==3.2.5
PySide6==6.11.0
PySide6_Addons==6.11.0
PySide6_Essentials==6.11.0
PySocks==1.7.1
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2026.1.post1
pywebview==6.2.1
PyYAML==6.0.3
pyzmq==27.1.0
qrcode==8.2
referencing==0.37.0
regex==2026.3.32
repath==0.9.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.7.2
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.15.3
seaborn==0.13.2
setuptools==80.10.2
shellingham==1.5.4
shiboken6==6.11.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.3
stack_data==0.6.3
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.12.1
tensorboard_data_server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
termcolor==3.2.0
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.4.0
ttach==0.0.3
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvloop==0.22.1
watchdog==6.0.0
wcwidth==0.6.0
Werkzeug==3.1.3
wheel==0.35.1
wrapt==2.0.1
yarl==1.22.0
